update: refactored calls to utils.Filemode to git_shell.Filemode
