+++ ggshield/scan/scannable.py
@@ -8,7 +8,7 @@ from typing import BinaryIO, Callable, List, Optional, Tuple
 import charset_normalizer
 from charset_normalizer import CharsetMatch
 
-from ggshield.core.utils import Filemode
+from ggshield.core.git_shell import Filemode
 
 
 logger = logging.getLogger(__name__)
